Refactor to NOT connect to device.
The way I was doing this before, where the app would connect to the
Kindle as if it were just another disk, only works with Android <=2.3.
Which means it doesn't work for Kindles newer that 1st or 2nd gen. :(

Android decided it was a security risk, so later versions of Android
only support MTP. I couldn't find a way to do MTP programmatically, so
I decided to just change the app to do everything on the computer.

Now, users need to copy the `com.deseretbook.bookshelf` folder from
their device to the Desktop of the computer in order for the app to
locate it. It's gross, but it works.